python/requirements.txt
python-pptx==1.0.2
requests==2.32.3
openai==1.59.0